Revised plotting infrastructure. Created calorimeter shower plotting functions. sVAE model running. Added README section.
