+++ configs/engine/calo_training.yaml
@@ -1,6 +1,6 @@
 # @package _group_
 learning_rate: 0.001
-n_batch_samples: 100
-n_epochs: 1
+n_batch_samples: 256
+n_epochs: 10
 weight_decay_factor: 0.0001
 momentum_coefficient: 0.5
