--- utils/plotProvider.py
@@ -1,111 +1,0 @@
-
-"""
-PlotProvider. Work in progress.
-
-Author: Eric Drechsler (eric_drechsler@sfu.ca)
-"""
-
-import os
-import torch
-import numpy as np
-
-from DiVAE import logging
-logger = logging.getLogger(__name__)
-
-from utils.helpers import plot_MNIST_output, plot_latent_space
-
-class PlotProvider(object):
-    def __init__(self,config_string="default",date_tag="000000", cfg=None):
-        self._config=cfg
-        
-        self.config_string=config_string
-        self.date_tag=date_tag
-
-        self._data_dimensions=None
-        pass
-
-    @property
-    def data_dimensions(self):
-        return self._data_dimensions
-
-    @data_dimensions.setter
-    def data_dimensions(self, dimensions):
-        self._data_dimensions=dimensions
-
-    def plot_generative_output(self, input_data):
-        #TODO this is a quick hack to use the same plotting function for input
-        #vs output and generated samples
-        if isinstance(input_data,list):
-        # plot_generative_output(output.detach(), n_samples=n_samples, output="./output/divae_mnist/rbm_samples/rbm_sampling_{0}.png".format(outstring))
-        # def generate_samples_svae(model, outstring=""):
-            outputs=[ out.detach() for out in input_data]
-            from utils.helpers import plot_calo_jet_generated
-            plot_calo_jet_generated(outputs, n_samples=5, output="./generated_{0}.png".format("test"))
-            return
-        else:   
-            n_split=input_data.size()[0]//2
-            in_data_1,in_data_2=torch.split(input_data,split_size_or_sections=int(n_split),dim=0)
-            in_data_1=in_data_1.detach().numpy()
-            in_data_2=in_data_2.detach().numpy()
-            
-            plot_MNIST_output(input_data=in_data_1, 
-                output_data=in_data_2,
-                n_samples=n_split, 
-                out_file="{0}/{2}_{1}_generated.png".format(self._config.output_path,self.config_string,self.date_tag))
-        pass
-
-    def plot(self, input_container):
-        logger.info("Plotting")
-        #the container with our output objects to plot
-        input_container.print()
-
-        if self._config.data.data_type.lower()=="mnist":
-            
-            #default plot method
-            plot_MNIST_output(input_data=input_container.input_data, 
-                output_data=input_container.output_data,
-                n_samples=self._config.n_plot_samples, 
-                out_file="{0}/{2}_reco_{1}.png".format(self._config.output_path,self.config_string,self.date_tag))
-            
-            #for visualisation purposes: plot latent spaces representation for certain models.
-            #See method for details.
-            #TODO this is a hacky switch
-            if not self._config.model.model_type=="cVAE" and not self._config.model.model_type=="DiVAE" and not self._config.model.model_type=="HiVAE":
-                plot_latent_space(zeta=input_container.zeta,
-                    label=input_container.labels, 
-                    out_file="{0}/{2}_latSpace_{1}".format(self._config.output_path,self.config_string,self.date_tag),
-                    dimensions=0)
-        
-        elif self._config.data.data_type.lower()=="calo":
-
-            from utils.helpers import plot_calo_images, plot_calo_image_sequence
-            
-            if self._config.model.model_type=="sVAE":
-                plot_calo_image_sequence(
-                    input_data=input_container.input_data, 
-                    output_data=input_container.outputs, 
-                    input_dimension=self.data_dimensions, 
-                    out_file="{0}/{2}_{1}.png".format(self._config.output_path,self.config_string,self.date_tag))
-            else:
-                test_loss, input_data, output_data, zetas, labels  = modelCreator.test()
-                plot_calo_images(input_data, output_data, output="{0}/{2}_reco_{1}.png".format(self._config.output_path,configString,date))
-        else:
-            raise Exception("Data type {0} unknown to PlotProvider".format(self._config.data.data_type))
-        
-        #TODO gif_output, plot_latent_space
-        #     gif_frames=[]
-        #     for epoch in range(1, self._config.engine.n_epochs+1):   
-        #         test_loss, input_data, output_data, zetas, labels  = modelCreator.test()
-        #         if self._config.create_gif:
-        #             if self._config.data.data_type=='calo':
-        #                 gif_frames.append(plot_calo_images(input_data, output_data, output="{0}/{2}_reco_{1}.png".format(self._config.output_path,configString,date),do_gif=True))
-        #             else:
-        #                 gif_frames.append(gif_output(input_data, output_data, epoch=epoch, max_epochs=self._config.engine.n_epochs, train_loss=train_loss,test_loss=test_loss))
-        #     if self._config.create_gif:
-        #         gif.save(gif_frames,"{0}/runs_{1}.gif".format(self._config.output_path,configString),duration=200)
-        pass
-
-if __name__=="__main__":
-    logger.debug("Testing PlotProvider Setup") 
-    model=PlotProvider()
-    logger.debug("Success")
